requirements.txt
langchain-ollama
langchain-core
langchain-community
langserve
fastapi
uvicorn